backend_api/app/requirements.txt
alembic==1.7.1
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
colorama==0.4.4
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.1
fastapi-jwt-auth==0.5.0
flake8==3.9.2
greenlet==1.1.1
h11==0.12.0
idna==3.2
importlib-resources==5.2.2
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
passlib==1.7.4
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==1.7.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
raven==6.10.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
sqlalchemy2-stubs==0.0.2a15
sqlmodel==0.0.4
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
zipp==3.5.0
tenacity==8.0.1
python-binance
google-auth